do not check automatic edits on */makefile.in. Handle cppad/config.h differently
git-svn-id: https://projects.coin-or.org/svn/CppAD/stable/20110101@1817 0e562018-b8fb-0310-a933-ecf073c7c197

+++ bin/commit.sh
@@ -68,10 +68,17 @@ then
 		exit 0
 	fi
 	abort="no"
+	config_changed="no"
+	if grep 'cppad/config\.h' bin/commit.1.$$ > /dev/null
+	then
+		config_changed="yes"
+		sed -i bin/commit.1.$$ -e '/cppad\/config\.h/d'
+	fi
 	list=`cat bin/commit.1.$$`
 	for file in $list
 	do
-		if [ -f "$file" ]
+		name=`echo $file | sed -e 's|.*/||'`
+		if [ -f "$file" ] && [ "$name" != "makefile.in" ]
 		then
 			sed -f bin/commit.sed $file > bin/commit.2.$$
 			if ! diff $file bin/commit.2.$$ > /dev/null
@@ -117,6 +124,11 @@ then
 	echo "------------------------------------"
      echo "chmod +x bin/commit.sh"
            chmod +x bin/commit.sh
+	if [ "$config_changed" = "yes" ]
+	then
+		echo "Not sure changes to cppad/config.h should be commited."
+		echo "You must include it yourself (by hand) if they should."
+	fi
 	exit 0
 fi
 # -----------------------------------------------------------------------
